% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_plot.R
\name{abs_pos}
\alias{abs_pos}
\title{abs_pos()}
\usage{
abs_pos(data.bin, genome = "hg38", true_loc = TRUE)
}
\description{
add columns of chrosome absolute start and end for "segName" of a data.frame
segName can be chr1-100037672-100038546 or chr1_100037672_100038546 format
}
