% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{align_Grange2bin}
\alias{align_Grange2bin}
\title{align_Grange2bin()}
\usage{
align_Grange2bin(bed.query, bed.subject)
}
\arguments{
\item{bed.query}{A bed file or a data frame of query regions}

\item{bed.subject}{A bed file or a data frame of subject regions}
}
\value{
A data frame of the same order as the reference bed file, with the corresponding bin ID for each query region.
}
\description{
align_Grange2bin()
}
