% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_plot.R
\name{annotate_gene}
\alias{annotate_gene}
\title{annotate_gene()}
\usage{
annotate_gene(gene_name, genome = "hg38", annotations = NULL)
}
\arguments{
\item{gene_name}{gene symbol}

\item{genome}{string of reference genome}

\item{annotations}{GRanges object}
}
\value{
a GRanges object of gene annotation
}
\description{
genomic annotation for gene symbol
}
