% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{atac_integration}
\alias{atac_integration}
\title{atac_integration()}
\usage{
atac_integration(
  obj1,
  obj2,
  peakwidth.lim = c(200, 10000),
  reduction = "rlsi",
  dims = 2:30
)
}
\description{
merge ATAC objects
}
