% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_filtMT.R
\name{bedtools_intersect}
\alias{bedtools_intersect}
\title{bedtools_intersect()}
\usage{
bedtools_intersect(
  stringsA,
  stringsB,
  wa = TRUE,
  sepA = c("-", "-"),
  sepB = c("-", "-")
)
}
\description{
bedtools_intersect()
}
