% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_bulkCNV_estimation.R
\name{bestCNVres.select}
\alias{bestCNVres.select}
\title{bestCNVres.select()}
\usage{
bestCNVres.select(inteCNV_res)
}
\arguments{
\item{inteCNV_res}{list of integer CNV result with length of N clusters}
}
\description{
selection the best integer CNV result from  multiple clusters based on the the Mean Squared Error (MSE) of integer CNV estimation to the mapping ratio
}
