% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_binClust.R
\name{binClust}
\alias{binClust}
\title{binClust()}
\usage{
binClust(
  inputMat,
  cellMeta,
  ref_group_names = NULL,
  chrs = NULL,
  delimit = "-",
  nFeature.min = 50,
  Center = TRUE,
  onRatio = TRUE,
  outdir = ".",
  window = 5,
  seu_resol = 0.8,
  outputFigures = TRUE,
  plt4peak = FALSE,
  OutMat = TRUE,
  Figure_name = NULL,
  noRefShow = FALSE,
  count_lim = NULL,
  Figure.width = 10,
  Figure.height = 6,
  cytoBand = NULL,
  global_norm = FALSE
)
}
\arguments{
\item{cellMeta}{data.frame of cell annotation, group information on the first column, cell IDs on rownames}

\item{ref_group_names}{the group name as reference}

\item{chrs}{analysis for specific chromosomes}

\item{cytoBandFile}{the reference annotation of chromosome location, BED format with columns of c("chrom","chromStart","chromEnd","name","gieStain")}
}
\description{
Process the input matrix into a bin count matrix if required, then calculate
the ratio relative to the reference, and perform PCA dimensional reduction and clustering
based on the ratio.
}
