% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_filtMT.R
\name{blacklist.rm}
\alias{blacklist.rm}
\title{blacklist.rm()}
\usage{
blacklist.rm(mat, blacklist, peak.sep = "-")
}
\arguments{
\item{mat}{count matrix with peaks on rows, cells on columns.
Peaks with "chrx-xxx-xxx" format were default separated with "-"}

\item{blacklist}{bed format file with path}
}
\description{
remove peaks in blacklist
}
