% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_filtMT.R
\name{cell_score}
\alias{cell_score}
\title{cell_score()}
\usage{
cell_score(
  data,
  ZeroProp.max = 0.5,
  byBin = TRUE,
  binWin = 20,
  outFigure = TRUE,
  outdir = "./",
  cell_meta = NULL,
  max.value = NULL
)
}
\description{
cell_score()
}
