% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_filtMT.R
\name{cell_VarScore}
\alias{cell_VarScore}
\title{cell_VarScore()}
\usage{
cell_VarScore(data, genomic_ref = NULL, Zscore.Label = 1)
}
\arguments{
\item{data}{peak- cell matrix (row:peak,column:cell)}

\item{genomic_ref}{Bed file of with genomic range for arms of chromosome (e.g. cytoBand_hg38.tsv)}
}
\description{
score cell based on the arm-levle weighted standard deviation
}
