% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{cellEst}
\alias{cellEst}
\title{cellEst()}
\usage{
cellEst(segratio.correct, clonalest, clonalest.Ref = NULL)
}
\arguments{
\item{segratio.correct:}{corrected ratio matrix}

\item{clonalest:}{clonal ploidy estimation}
}
\value{
cell-integerCN matrix
}
\description{
integer CN estimation at cell level
}
