% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{cellintegerCN}
\alias{cellintegerCN}
\title{cellintegerCN()}
\usage{
cellintegerCN(cellratioID, finalres, refClone = NULL)
}
\arguments{
\item{cellratioID:}{cell ratio}

\item{finalres:}{clonal estimation}
}
\value{
cell-bin matrix: integerCN
}
\description{
cell level integerCN estimation
}
