% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figurePlot.R
\name{cellLineSeg_v3}
\alias{cellLineSeg_v3}
\title{cellLineSeg_v3()}
\usage{
cellLineSeg_v3(
  binRatio.df = NULL,
  integerCNV.df = NULL,
  CNVres = NULL,
  integerCNV = NULL,
  outdir,
  fname,
  ylim = NULL,
  genome = "hg38",
  ylab = "Relative CN",
  plot_seg = TRUE,
  color_seg_gradient = TRUE,
  seg.color = "#0d3b66",
  ggarrange_width = c(10, 5),
  height = 2,
  width = 15,
  plot_dots = TRUE,
  color_dot = FALSE,
  plot_hist = TRUE,
  color_hist_gradient = TRUE,
  hist_color = NULL,
  color_limit = c(1, 6),
  label_gene = NULL,
  label_gene_color = "blue",
  annotations = NULL,
  outPlot = TRUE,
  length.out = 100,
  value.bin = "ratio",
  value.segment = "SegMean",
  label_CN = TRUE
)
}
\arguments{
\item{CNVres}{list out put from dataProcess()function}

\item{integerCNV}{list output from doCNV1_v2() function}
}
\description{
plot dots and histogram of segment ratio at true chromosome location,
remove bad segment from histogram and hist based on the true genomic length of segments,set a global color
}
