% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_atac2cnv.R
\name{celloutput}
\alias{celloutput}
\title{celloutput()}
\usage{
celloutput(
  cells.obs,
  cluster1,
  cluster2,
  clone_res,
  mtx_bin,
  CNest.ref,
  min_cells_in_group = 20,
  penalty.lim = c(0.5, 1.5),
  seg_method = "PELT",
  FiltSeg = TRUE,
  SegLen_min = 2e+06,
  SegSize_min = 10,
  seg.count.lim = 80,
  cytoBandFile = NULL,
  outdir = "./",
  minCN.frac = 0.01,
  seg_dat_ref,
  correct_by_dist = FALSE,
  delt_lim = 0.1,
  segValue_method = "median",
  p.adj_cutoff = 0.05,
  ratio_diff_cutoff = NULL,
  ...
)
}
\arguments{
\item{cells.obs}{the input cells}

\item{cluster1}{group annotation for cells.obs}

\item{cluster2}{updated group annotation for cells.obs}
}
\description{
celloutput()
}
