% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{cellRatio_corrected}
\alias{cellRatio_corrected}
\title{cellRatio_corrected()}
\usage{
cellRatio_corrected(cellratioID, finalres, refClone = NULL)
}
\arguments{
\item{cellratioID:}{initial cell ratio}

\item{finalres:}{clonal estimation}
}
\value{
cell-bin matrix: corrected Ratio
}
\description{
cell level corrected ratio
}
