% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_atac2cnv.R
\name{clone_scoring}
\alias{clone_scoring}
\title{clone_scoring()}
\usage{
clone_scoring(
  cloneRes,
  scaleFactor = 1,
  y_true.term = "ratio",
  y_pred.term = "relativeCN",
  CNbase = cloneRes$CNest,
  ploidy.ref = NULL
)
}
\arguments{
\item{cloneRes}{the output list of ploidyRefine(), i.e. clonal_res}
}
\description{
clone_scoring()
}
