% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.subclonality.function.R
\name{cloneRatio}
\alias{cloneRatio}
\title{cloneRatio()}
\usage{
cloneRatio(ratiodata, bindata, chr)
}
\arguments{
\item{ratiodata:}{initial clone list}
}
\value{
all ratio value frm all clone and genomic binID
}
\description{
collect overall ratio at clone level
}
