% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_bulkCNV_estimation.R
\name{CNbaseline.fill}
\alias{CNbaseline.fill}
\title{CNbaseline.fill()}
\usage{
CNbaseline.fill(CNbase, delt = NULL)
}
\arguments{
\item{CNbase}{data.frame with columns of 'ratio' and 'CN'}
}
\description{
fill the blank CN level based on known delt and CNs
}
