% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_find_subclones.R
\name{CNratio_peak}
\alias{CNratio_peak}
\title{CNratio_peak()}
\usage{
CNratio_peak(
  mt,
  cell_ano,
  ref.value = NULL,
  colname_ref = "adjacent",
  Norm_by_ref = FALSE
)
}
\arguments{
\item{mt}{combined single cell matrix}

\item{cell_ano}{a data.frame has same number of cells on row, cell ID (first column) and sub-cluster annotation (second column) on columns
This function will convert single-cell matrix to cluster-level matrix first based on sub-cluster annotation}

\item{ref.value}{numeric data.frame with the length of nrow(mat)}

\item{colname_ref}{the name of group taken as reference.}
}
\description{
calculate peak-cluster level CN ratio
}
