% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_atac2cnv.R
\name{CNratioInfer}
\alias{CNratioInfer}
\title{CNratioInfer()}
\usage{
CNratioInfer(
  input_matrix,
  inputcellano,
  cytoBand,
  outdir = NULL,
  group2bulkby = "subCluster",
  min_cells_in_group = 1,
  normallabel = "reference",
  cells.ref = NULL,
  seg_method = "PELT",
  seg.count.lim = 80,
  penalty = 1,
  segValue_method = "density",
  FiltSeg = TRUE,
  SegLen_min = 1e+06,
  SegSize_min = 5,
  plot_ylim = NULL,
  outFigure = TRUE,
  outFigureName = "CNRatio_subCluster.pdf",
  color_dot = FALSE,
  output_ref.value = TRUE,
  outRefValueName = "meanCount_Ref.rds",
  outRDS = TRUE,
  RDSname = "segScore_median.rds"
)
}
\description{
infer CN ratio for group and segmentation
}
