% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_atac2cnv.R
\name{CNratioInfer.sc}
\alias{CNratioInfer.sc}
\title{CNratioInfer.sc()}
\usage{
CNratioInfer.sc(
  input_matrix,
  cell.name,
  inputcellano,
  cytoBandFile,
  group2bulkby = "subCluster",
  normallabel = "reference",
  seg.ref = NULL,
  seg_method = "robseg",
  segValue.method = "density",
  outdir = ".",
  scale_value = TRUE,
  prop_zero = NULL,
  outPlot = TRUE,
  outFigureName.prefix = "segRatio_sc",
  add_yline = NULL
)
}
\arguments{
\item{prop.zero}{numeric vector with the same length of rows of input_matrix.}
}
\description{
infer CN ratio for individual single cell
}
