% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{CNupdate}
\alias{CNupdate}
\title{CNupdate()}
\usage{
CNupdate(df_seg_C2, initialCN)
}
\arguments{
\item{df_seg_C2:}{segmentation data}

\item{initialCN:}{the initial estimation from ABSOLUTE model}
}
\value{
the likelihood for each ploidy estimation
}
\description{
estimating the corrected ratio baseline and distance used in PloidyCorrect
}
