% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_atac2cnv.R
\name{CNV.align}
\alias{CNV.align}
\title{CNV.align()}
\usage{
CNV.align(
  df_seg_C1,
  initialCN,
  subclone.rawBase = NULL,
  rato.correct = TRUE,
  align_by = "SegMean"
)
}
\arguments{
\item{rato.correct}{subclone ratio correction}
}
\description{
align CNVs based on the baseline
}
