% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_bulkCNV_estimation.R
\name{CNV_esti}
\alias{CNV_esti}
\title{CNV_esti()}
\usage{
CNV_esti(
  outdir = ".",
  segScore_file = "segScore_median.rds",
  filt_seg = TRUE,
  offset.adjust = TRUE,
  length_seg_cutoff = 1e+06,
  genome = "hg38",
  true_loc = TRUE,
  outplot_name = NULL,
  delta.min = 0.4,
  delta.max = 0.7,
  outputFigure = TRUE
)
}
\arguments{
\item{segScore_file}{the file with segment ratio of each bin for each cell}

\item{offset.adjust}{Adjust the ratio to nearest integer.}

\item{length_seg_cutoff}{if filt_seg=TRUE, set the quantile threshold length of segments}

\item{genome}{the reference genome for the genomic location annotation}
}
\description{
CNV_esti()
}
