% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_bulkCNV_estimation.R
\name{CNV_esti_ini}
\alias{CNV_esti_ini}
\title{CNV_esti_ini()}
\usage{
CNV_esti_ini(
  outdir = ".",
  segScore_file = "segScore_median.rds",
  filt_seg = TRUE,
  size_seg_cutoff = 5,
  length_seg_cutoff = 1e+06,
  segValue_method = "median",
  genome = "hg38",
  true_loc = TRUE,
  outplot_name = NULL,
  outputFigure = TRUE
)
}
\description{
estimate initial clonal-level CNVs
}
