% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_utils.R
\name{cnv_score}
\alias{cnv_score}
\title{cnv_score()}
\usage{
cnv_score(
  mat,
  diploid.base = NULL,
  diploid.base.sd = NULL,
  Var.cutoff = NULL,
  ref_cells = NULL,
  plot_out = TRUE,
  outdir = "./",
  scaleFactor = 1
)
}
\arguments{
\item{mat}{CNV matrix (cells in columns, genes or peaks in rows)}

\item{diploid.base}{diploid value}
}
\description{
Score cells based on their CNV profile, then impute malignant cells.
}
