% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_atac2cnv.R
\name{combine_res_chr}
\alias{combine_res_chr}
\title{combine_res_chr()}
\usage{
combine_res_chr(
  data.dir,
  mtx_seg,
  chrs,
  df_meta_slim,
  groups_col = "subclone",
  outdir = ".",
  outFile = TRUE,
  estract_by = "SegMean"
)
}
\arguments{
\item{data.dir}{contains chr folders}

\item{mtx_seg}{NA matrix with colname with cell and rowname with peak}
}
\value{
matrix of peak-cell seg.ratio combined different cluster of cells from each chromosome
}
\description{
combine the segment ratio result from each chromosome
}
