% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_find_subclones.R
\name{compare2RefSeg}
\alias{compare2RefSeg}
\title{compare2RefSeg()}
\usage{
compare2RefSeg(
  df_ref,
  df_query,
  df_binRatio = df_query,
  delta_base = NULL,
  diff_prop_lim = 0.01,
  reCalcuCNA = FALSE
)
}
\arguments{
\item{df_ref}{data.frame with c("binID","ratio_map","integerCNV") columns}

\item{df_query}{data.frame with "binID", "SegMean" columns}
}
\description{
the proportion of peaks (or bins) with different integerCNV  between df_query and df_ref
}
