% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_utils.R
\name{Correct_by_length}
\alias{Correct_by_length}
\title{Correct_by_length()}
\usage{
Correct_by_length(mat, unit = 1000)
}
\arguments{
\item{mat}{peak-cell count matrix, rowname is in "chrx-xxx-xxx" format}

\item{unit}{per base, kb or Mb, default is 1000b (kb).}
}
\description{
correct count by peak length
}
