% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_plot.R
\name{correlat_plot}
\alias{correlat_plot}
\title{correlat_plot()}
\usage{
correlat_plot(
  seu_obj = NULL,
  mt_atac = NULL,
  mt_rna = NULL,
  cnv_rna = NULL,
  assay.atac = "peaks",
  assay.rna = "RNA",
  cellMeta,
  CN_res,
  FiltSegbyLen = TRUE,
  segLen_lim = 1e+06,
  outdir = "./",
  dotPlot4Seg = TRUE,
  label_term = NULL
)
}
\arguments{
\item{seu_obj}{Seurat object}

\item{mt_atac}{(Optional)scATAC count matrix (normalized), peaks on row and cells on column}

\item{mt_rna}{(Optional)scRNA count matrix (normalized), genes on row and cells on column}

\item{cellMeta}{data.frame with cell group annotation on the second column, cellID on the rowname and first column}

\item{CN_res}{final CNV result output by runEiCNVs()}
}
\description{
correlat_plot()
}
