% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_find_subclones.R
\name{count2ratio}
\alias{count2ratio}
\title{count2ratio()}
\usage{
count2ratio(
  mt_obs,
  mt_ref,
  colname_ref = "adjacent",
  win = 10,
  chr_select = NULL
)
}
\arguments{
\item{mt_obs}{and mt_ref: peak-cell count matrix of tumor and normal reference, which rows should be the same}

\item{colname_ref}{the name of group taken as reference.}

\item{win}{The number of features to be combined into one bin.}

\item{chr_select}{This is only needed when performed on a specific chromosome.}
}
\value{
cell_info a data.frame with cell ID and sub-cluster annotation on columns
}
\description{
count2ratio()
}
