% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeaCNV_main.R
\name{CreateTeaCNVObject}
\alias{CreateTeaCNVObject}
\title{CreateTeaCNVObject()}
\usage{
CreateTeaCNVObject(
  input,
  annotationFile,
  ref_group_names,
  assay = "ATAC",
  delim = "\\t",
  genome = "hg38",
  blacklistFile = NULL,
  cytoBandFile = NULL,
  ChrRemove = c("chrX", "chrY", "chrM"),
  FiltCell = TRUE,
  FiltCell_by = "quantile",
  CellnCount_quant = c(0.05, 0.95),
  CellnFeature_quant = c(0.05, 0.95),
  cellproplim = 0.05,
  count_lim = 4,
  prop_reset = TRUE
)
}
\arguments{
\item{FiltCell_by}{Filtering method if FiltCell=TRUE, c("Zscore","density","quantile"),default is "Zscore".}
}
\description{
CreateTeaCNVObject()
}
