% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{dataProcess}
\alias{dataProcess}
\title{dataProcess()}
\usage{
dataProcess(CNVdata, method = "median")
}
\arguments{
\item{method}{"mean", "median" or "density" for calculate the value of segment.}
}
\description{
dataProcess()
}
