% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{denoise_via_global_mean}
\alias{denoise_via_global_mean}
\title{denoise_via_global_mean()}
\usage{
denoise_via_global_mean(mat_obj, sd_amplifier = 1)
}
\arguments{
\item{mat_obj}{cells in columns}
}
\description{
denoise_via_global_mean()
}
