% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_InformationSegs.R
\name{densityPlot_segs}
\alias{densityPlot_segs}
\title{densityPlot_segs()}
\usage{
densityPlot_segs(clonalest_ls, seg.bed, cols_Palette = NULL, xlim = NULL)
}
\arguments{
\item{clonalest_ls}{list of clonal-CNVs results}
}
\value{
list of plots
}
\description{
densityPlot_segs()
}
