% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_segmentation_comb.R
\name{DEsegs}
\alias{DEsegs}
\title{DEsegs()}
\usage{
DEsegs(
  df,
  group1 = "refC",
  group2 = "obsC",
  wide2long_by = "binID",
  p.adj_cutoff = 0.05,
  ratio_diff_cutoff = NULL,
  na_rm = TRUE
)
}
\arguments{
\item{data.frame}{with columns: bin name, segment name, ratio of group1, ratio of group2}
}
\value{
mode=2 in the output data.frame means the segment with differential mean ratio between groups.
}
\description{
Test the difference in ratio values of all bins on each segment between the two groups.
}
