% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{.find_max_density}
\alias{.find_max_density}
\title{.find_max_density()}
\usage{
.find_max_density(counts_analysis, log_scale = T)
}
\arguments{
\item{counts_analysis}{vector of data}

\item{log_scale}{if TRUE, density distribution based on log transformed value}
}
\value{
the value at the max density
}
\description{
.find_max_density()
}
