% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_pseudoBulk_mat.R
\name{.get_ref_density}
\alias{.get_ref_density}
\title{.get_ref_density()}
\usage{
.get_ref_density(matrix, ref_indice, zero.rm = FALSE)
}
\arguments{
\item{matrix}{input matrix, group in columns (cells)}

\item{ref_indice}{numeric indices for group}
}
\description{
.get_ref_density()
}
\keyword{function}
\keyword{internal}
