% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_pseudoBulk_mat.R, R/mydataProcess.R
\name{.get_ref_mean}
\alias{.get_ref_mean}
\title{.get_ref_mean()}
\usage{
.get_ref_mean(matrix, ref_indice, zero.rm = FALSE)

.get_ref_mean(matrix, ref_indice, zero.rm = FALSE)
}
\arguments{
\item{matrix}{input matrix, group in columns (cells)}

\item{ref_indice}{numeric indices for group}
}
\description{
.get_ref_mean()
}
\keyword{function}
\keyword{internal}
