% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{.smooth_by_harmonic_mean}
\alias{.smooth_by_harmonic_mean}
\title{.smooth_by_harmonic_mean()}
\usage{
.smooth_by_harmonic_mean(Mat, wl = NULL, step = 1, sep_by = "-")
}
\description{
calculate harmonic mean with weighted by the number of non-zero within the window
wl is the window length, which is the number of peaks(default wl=20)
}
\keyword{internal}
