% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{.smooth_by_maxDensity}
\alias{.smooth_by_maxDensity}
\title{.smooth_by_maxDensity()}
\usage{
.smooth_by_maxDensity(matrix, wl = NULL, cores = 1, log_scale)
}
\arguments{
\item{matrix}{counts matrix with peaks(rows) and cells (columns).The row names should be with "chr1_247368351_247368851" format.}

\item{wl}{window length, which is the number of neighbor features to use for smoothing.}

\item{cores}{the number of cores if doParallel. Default is 1.}
}
\description{
To smooth for sparse, large scATAC data per chromosome, identify the max density distributions within a window,
and take the value under the max density to replace all the values within this window region.
}
\keyword{internal}
