% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{.smooth_by_sum}
\alias{.smooth_by_sum}
\title{.smooth_by_sum()}
\usage{
.smooth_by_sum(matrix, wl = NULL, cores = 1)
}
\description{
calculate sum within the window
wl is the window length, which is the number of peaks(default wl=200)
}
\keyword{internal}
