% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{.smooth_by_sum2bin}
\alias{.smooth_by_sum2bin}
\title{.smooth_by_sum2bin()}
\usage{
.smooth_by_sum2bin(Mat, wl = NULL, sep_by = "-")
}
\description{
calculate sum within the window,
and shrink the raw matrix (rows) by combining the peaks to bins decided by the start and end site of peaks within the window
wl is the window length, which is the number of peaks(default wl=200)
}
\keyword{internal}
