% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{.smooth_by_weighted_mean}
\alias{.smooth_by_weighted_mean}
\title{.smooth_by_weighted_mean()}
\usage{
.smooth_by_weighted_mean(Mat, wl = NULL, step = 1, sep_by = "-")

.smooth_by_weighted_mean(Mat, wl = NULL, step = 1, sep_by = "-")
}
\description{
calculate mean with weighted by the length of peaks within the window
wl is the window length, which is the number of peaks(default wl=20)

calculate mean with weighted by the length of peaks within the window
wl is the window length, which is the number of peaks(default wl=20)
}
\keyword{internal}
