% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{.smooth_by_windowMean}
\alias{.smooth_by_windowMean}
\title{.smooth_by_windowMean()}
\usage{
.smooth_by_windowMean(Mat, wl = NULL, step = 1, zero.rm = FALSE, sep_by = "-")
}
\arguments{
\item{wl}{window length, which is the number of neighbor features to use for smoothing.
@param step the moving step of a sliding window}
}
\description{
calculate MEAN within the window,
and convert the raw matrix (rows) by combining the peaks to bins decided by the start and end site of peaks within a window step
}
\keyword{Example:}
\keyword{Mat=matrix(seq(1:150),nrow=30)}
\keyword{internal}
\keyword{rownames(Mat)=paste(rep("chr1",30),seq(1:30),seq(201,230),sep="_")}
