% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{.smooth_by_windowMedian}
\alias{.smooth_by_windowMedian}
\title{.smooth_by_windowMedian()}
\usage{
.smooth_by_windowMedian(Mat, wl = NULL, step = 1, sep_by = "-")
}
\description{
.smooth_by_windowMedian()
}
