% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeaCNV_main.R
\name{EffectCorrectByRef}
\alias{EffectCorrectByRef}
\title{EffectCorrectByRef()}
\usage{
EffectCorrectByRef(
  mat,
  cell_anno,
  sampleID_normal,
  NormalTypeList,
  sampleIDs_to_correct,
  Ncells.min = 10
)
}
\arguments{
\item{cell_anno}{data.frame of cell annotation, containing the cell name(row.name),
the cell group classification(first column) and the sample ID (second column), tab-delimited}

\item{sampleID_normal}{Normal samples in the second column of cell_anno}

\item{NormalTypeList}{cell groups in the first column of cell_anno}
}
\description{
EffectCorrectByRef()
}
