% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_filtMT.R
\name{filt_peak_perChr}
\alias{filt_peak_perChr}
\title{filt_peak_perChr()}
\usage{
filt_peak_perChr(mat, zscore.lim = 2, split_by = "-")
}
\arguments{
\item{mat}{peak-cell count matrix}

\item{zscore.lim}{the Zscore cutoff value for distribution of peak total count}

\item{split_by}{the delimiter of peak name, e.g. "chrx-xxx-xxx"}
}
\description{
Remove peaks with abnormally high total count in all cells per chromosome
}
