% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_find_subclones.R
\name{filt_peaks_bySum}
\alias{filt_peaks_bySum}
\title{filt_peaks_bySum()}
\usage{
filt_peaks_bySum(mat, rows, zscore.lim = 2)
}
\arguments{
\item{mat}{matrix with peaks on row, cells on clumn}
}
\description{
filt_peaks_bySum()
}
