% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_filtMT.R
\name{FiltCell.obj}
\alias{FiltCell.obj}
\title{FiltCell.obj()}
\usage{
FiltCell.obj(
  object,
  assay = "peaks",
  cell_meta = NULL,
  filt_by = paste0(c("nCount_", "nFeature_"), assay),
  filt_per_group = NULL,
  by_density = TRUE,
  by_quantile = FALSE,
  by_zscore.x = TRUE,
  zscore.min = -1,
  nCount_quantile_lim = c(0.05, 0.95),
  nFeature_quantile_lim = c(0.05, 0.95),
  log_ = TRUE
)
}
\value{
list of filtered cells and plots
}
\description{
FiltCell.obj()
}
