% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{FinalCN}
\alias{FinalCN}
\title{FinalCN()}
\usage{
FinalCN(initialCN1, disupdate, df_seg_C1)
}
\arguments{
\item{initialCN1:}{the estimation of ratio baseline}

\item{disupdate:}{all possible distance}

\item{df_seg_C1:}{segmentation data}
}
\value{
CNest: the final integerCN baseline
}
\description{
determine the final integerCN
}
