% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_InformationSegs.R
\name{find_infoSeg}
\alias{find_infoSeg}
\title{find_infoSeg()}
\usage{
find_infoSeg(
  clonalest_ls,
  seg_len_min = 2e+06,
  doPlot = TRUE,
  outdir = "./",
  cols_Palette = NULL,
  p.adj_cutoff = 0.05,
  ratio_diff_cutoff = NULL
)
}
\arguments{
\item{clonalest_ls}{list of clonal-CNVs results}
}
\description{
find_infoSeg()
}
