% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_find_subclones.R
\name{find_mean}
\alias{find_mean}
\title{find_mean()}
\usage{
find_mean(mt, cutoff, order_col = "l2h", zero.rm = TRUE)
}
\arguments{
\item{mt}{matrix with cells in column and features in row}

\item{order_col}{order columns from low to high (l2h) or high to low (h2l)}
}
\description{
Sort the cells according to the expression level from low to high,
increase the number of cells from the first one until the median value of the
row mean of the selected cells reaches the threshold, and return the selected cells
}
