% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{findccloneNum}
\alias{findccloneNum}
\title{findccloneNum()}
\usage{
findccloneNum(CNdata, n = 10)
}
\arguments{
\item{CNdata:}{cell-bin integerCN matrix}

\item{n:}{the maximumal number of clusters}
}
\description{
the optimal cluster number
}
