% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_find_subclones.R
\name{findClone.SortCells}
\alias{findClone.SortCells}
\title{findClone.SortCells()}
\usage{
findClone.SortCells(
  mtx_sub,
  df_seg_baseline,
  ref.value,
  cytoBand = NULL,
  Ncs = 20,
  er = 0.005,
  outdir = "."
)
}
\description{
findClone.SortCells()
}
