% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_find_subclones.R
\name{findRefClone}
\alias{findRefClone}
\title{findRefClone()}
\usage{
findRefClone(
  mat,
  cell_metadata,
  seg.method = "gaussian",
  seg.penalty = 0.4,
  outplot_path = "./"
)
}
\arguments{
\item{mat}{input matrix with cell barcodes on columns}

\item{cell_metadata}{data.frame with cell barcodes as rownames and cell annotation on the fist column}
}
\description{
find Reference Clone based On Clonal Segmentation for Ratio
}
