% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{gen_bin_cell_mtx}
\alias{gen_bin_cell_mtx}
\title{gen_bin_cell_mtx()}
\usage{
gen_bin_cell_mtx(
  mat_obj,
  bin_size = NULL,
  window_step = NULL,
  chrom_size = NULL,
  doFixBin = FALSE,
  sep_by = "-"
)
}
\arguments{
\item{mat_obj}{A count matrix of peak (or fragment) by cell}

\item{bin_size}{each bin size, i.e., sliding window size}

\item{window_step}{the step of the sliding window}

\item{chrom_size}{the data.frame of chromosome size with two columns: chr and size}
}
\description{
Generate bin by cell matrix for scATAC-seq data from peak by cell matrix
The matrix also can be generated by https://github.com/seasoncloud/Basic_CNV_SNV/blob/main/scripts/Gen_bin_cell_atac.R using "fragments.tsv.gz" file

if do fixed window bin, the bin_bed can be generate by "bedtools makewindows -g chrom.sizes -w 1000000 > genome.1Mb.bed"
or GenomicRanges::tileGenome() function with tilewidth=bin_size for chrosome size
}
