% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_pseudoBulk_mat.R
\name{gen_pseudo_mat}
\alias{gen_pseudo_mat}
\title{gen_pseudo_mat()}
\usage{
gen_pseudo_mat(
  obj,
  randomseeds,
  Neighbors = 100,
  assay = "RNA",
  expMatrix = obj[[assay]]@data,
  rowValue = "rowSums"
)
}
\arguments{
\item{obj}{seurat object, need to run FindNeighbors previously.}

\item{expMatrix}{sparse Matrix of class "dgCMatrix" for expression}
}
\description{
generate pseudobulk expression for seurat object based on N neighbors RNA expression
}
