% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.check.R
\name{GenomeHetFilter}
\alias{GenomeHetFilter}
\title{GenomeHetFilter()}
\usage{
GenomeHetFilter(
  obs,
  mode.res,
  max.non.clonal,
  max.neg.genome,
  Q,
  verbose = FALSE
)
}
\description{
GenomeHetFilter()
}
