% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.subclonality.function.R
\name{genomicPartition}
\alias{genomicPartition}
\title{genomicPartition()}
\usage{
genomicPartition(chrratio)
}
\arguments{
\item{chrratio:}{initial clone list}
}
\value{
genomic region partitioning
}
\description{
genome bin partition based on overall ratio
}
