% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.subclonality.function.R
\name{genomicRegion}
\alias{genomicRegion}
\title{genomicRegion()}
\usage{
genomicRegion(ratiodata, chr)
}
\arguments{
\item{ratiodata:}{initial clone list}
}
\value{
all genomic binID
}
\description{
collect all genome region
}
