% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{get_seg_data}
\alias{get_seg_data}
\title{get_seg_data()}
\usage{
get_seg_data(mtx_seg, outdir, Project, level = "segment", plot = TRUE)
}
\arguments{
\item{mtx_seg}{a list with 'data_matrix' bin-cell matrix; 'changepoint_matrix' with 1 means the segment end position; 'Ncells_groups',cell group if analysis based on groups (NA default)}

\item{outdir}{path for output file}

\item{Project}{Project for the file name output}

\item{level}{the output file is "bin" level or "segment" (default) level}
}
\description{
get segment dataframe from the bin-cell matrix after segmentation with changepoint information
}
