% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optEstimation.R
\name{GetSegQzPost}
\alias{GetSegQzPost}
\title{GetSegQzPost()}
\usage{
GetSegQzPost(
  d,
  sigma,
  w,
  comb,
  lambda.qz.res,
  sigma.h,
  theta.qz = NA,
  comb.s = 1
)
}
\description{
GetSegQzPost()
}
