% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optEstimation.R
\name{GetThetaQzPost}
\alias{GetThetaQzPost}
\title{GetThetaQzPost()}
\usage{
GetThetaQzPost(
  d,
  sigma,
  w,
  comb,
  sigma.h,
  pi_theta_qz,
  theta.qz = NA,
  comb.s = 1
)
}
\description{
GetThetaQzPost()
}
