% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_utils.R
\name{global_offset_correct}
\alias{global_offset_correct}
\title{global_offset_correct()}
\usage{
global_offset_correct(dat, dat.ref = NULL)
}
\arguments{
\item{dat}{a numeric vector of data to be correction}

\item{breaks}{a integer number for hist()}
}
\value{
the global CN ratio offset
}
\description{
Correct the global CN ratio offset by the distance between 2 times the maximum peak of the histogram and the nearest integer CNV.
if provided a reference correction, then use the maximum peak of the reference data as target to calculate the distance.
}
