% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figurePlot.R
\name{heatmap4peakMt}
\alias{heatmap4peakMt}
\title{heatmap4peakMt()}
\usage{
heatmap4peakMt(
  mat,
  meta_info = NULL,
  max_lim = NULL,
  sep_by = "-",
  outdir = "./",
  value.type = "count",
  max.value = NULL,
  color_bars = "discrete",
  n_breaks = 5,
  width = 10,
  height = 6,
  col_list = NULL,
  fileout_name = NULL,
  clust_rows = FALSE,
  clustering_method_rows = "complete",
  legend_direction = "vertical",
  legend_titles = NULL,
  heatmap_legend_side = "right",
  show_legend_row = TRUE,
  column.title = "Genomic Region",
  draw_normal = FALSE,
  ref_group_names = NULL,
  custom_colors = NULL,
  device = "png"
)
}
\arguments{
\item{mat}{matrix with peak in row and cells in columns, rownames pattern "chr-xxx-xxx"}

\item{sep_by}{a symbol for the row names separation}
}
\description{
heatmap for the 'mat' ordered by row of meta_info
}
