% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figurePlot.R
\name{HeatmapPlot}
\alias{HeatmapPlot}
\title{HeatmapPlot()}
\usage{
HeatmapPlot(
  dat,
  plotDir,
  type = "any",
  fname = NULL,
  clust = FALSE,
  clustering_method_rows = "complete",
  row_anno = NULL,
  width = 20,
  height = 9,
  show_row_names = FALSE,
  left_anno_col = NULL,
  color_bars = "discrete",
  n_breaks = 6,
  column.title = "Genomic Region",
  legend.direction = "vertical",
  legend_side = "right",
  legend_titles = NULL,
  show_legend_row = TRUE,
  draw_normal = FALSE,
  ref_group_names = NULL,
  max.legend.value = NULL,
  custom_colors = NULL,
  device = "png"
)
}
\arguments{
\item{dat}{ratio or CNV matrix with columns are cells and rows are features,first three columns are c("chr","start","end"), other columns are cells}

\item{row_anno}{data.frame with group annotation in columns and cellID as rownames}

\item{ref_group_names}{group name in the first column of row_anno}

\item{custom_colors}{list of paramters for color values}
}
\description{
HeatmapPlot()
}
