% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figurePlot.R
\name{hist_custom}
\alias{hist_custom}
\title{hist_custom()}
\usage{
hist_custom(
  data,
  x,
  color_hist = FALSE,
  color_col = NULL,
  row_filt_by = NULL,
  add_density = TRUE,
  scale_factor = 1,
  adjust = 1,
  log_y = FALSE,
  plotbreak = FALSE,
  y_cut = NULL,
  max_lim_y = NULL,
  stepby = NULL,
  cols_fill = NA,
  title_x = NULL,
  title_y = "density",
  binwidth = 0.1,
  xlim = NA,
  legend.position = "none",
  legend.title = "",
  title = "",
  color_hist_by_value = TRUE
)
}
\description{
hist_custom()
}
