% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figurePlot.R, R/ploidy.correction.v2.R
\name{hist_seg}
\alias{hist_seg}
\title{hist_seg()}
\usage{
hist_seg(
  df_seg_C1,
  integerCNV,
  ylim = NULL,
  color_limit = c(0, 3),
  cluster,
  length.out = 80
)

hist_seg(
  df_seg_C1,
  integerCNV,
  ylim = NULL,
  color_limit = c(0, 3),
  cluster,
  length.out = 80
)
}
\description{
hist_seg()

hist_seg()
}
