% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{hist_seg1}
\alias{hist_seg1}
\title{hist_seg1()}
\usage{
hist_seg1(
  df_seg_C1,
  integerCNV,
  ylim = NULL,
  color_limit = c(0, 3),
  cluster,
  length.out = 100
)
}
\description{
hist_seg1()
}
