% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{hist_seg2}
\alias{hist_seg2}
\title{hist_seg2()}
\usage{
hist_seg2(
  clusterEst,
  ylim = NULL,
  color_limit = c(0.5, 4),
  cluster = NULL,
  length.out = 100,
  color_hist_gradient = FALSE,
  hist_color = NULL,
  gtitle = NULL
)
}
\arguments{
\item{clusterEst:the}{colonal estimation}
}
\description{
plot the barplot of ratio as well as corresponding integer CN
}
