% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_utils.R
\name{infer_intCNV}
\alias{infer_intCNV}
\title{infer_intCNV()}
\usage{
infer_intCNV(
  CNVdata,
  SegLen_cutoff = NULL,
  shift_correc = TRUE,
  delta_cutoff = 0.4,
  delta_lim = c(0.4, 0.7)
)
}
\arguments{
\item{CNVdata}{input data.frame, bins on row, The columns should be 'chrom, start, end,ratio,breakpoint'
segName chrom                    binID Chromosome     Start       End binRatio
1 chr1_144551589_169368392     1 chr1-155261887-155262788       chr1 155261887 155262788 2.647409
2 chr1_144551589_169368392     1 chr1-155250380-155251334       chr1 155250380 155251334 4.546832}

\item{SegLen_cutoff}{quantile of segment length to be filtered out}
}
\description{
infer_intCNV()
}
