% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_atac2cnv.R
\name{ls.colummn2mtx}
\alias{ls.colummn2mtx}
\title{ls.colummn2mtx()}
\usage{
ls.colummn2mtx(res.ls, metadata, extract_by = "integerCNV")
}
\arguments{
\item{res.ls}{list of resuts of multiple clusters}
}
\description{
the names of res.ls should in first column of metadata
}
