% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{map2chrLocatio}
\alias{map2chrLocatio}
\title{map2chrLocatio()}
\usage{
map2chrLocatio(mat_obj, cytoBand = NULL, merge_cyto = TRUE)
}
\arguments{
\item{mat_obj}{matrix with 'chr?-startCoordinate-endCoordinate' as rownames,cell Id as colnames}
}
\value{
res, list of GRanges subject_syto, subject_observe, subject_overlap, and overlap row index
}
\description{
map to chromosome location,return a GRanges subject
}
