% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmode.R
\name{MargModeFinder}
\alias{MargModeFinder}
\title{MargModeFinder()}
\usage{
MargModeFinder(
  obs,
  dom2,
  Q,
  lambda.qz.res,
  pz,
  sigma.h,
  tau_dom,
  d.res = 0.025,
  verbose = FALSE
)
}
\description{
MargModeFinder()
}
