% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.subclonality.function.R
\name{mergeClones}
\alias{mergeClones}
\title{mergeClones()}
\usage{
mergeClones(
  ratiodata,
  segdata,
  doPlot = FALSE,
  outdir = "./",
  Zscore.cutoff = 1.28,
  p.adj.cutoff = 0.1,
  p.cutoff = NULL
)
}
\arguments{
\item{ratiodata}{list of "step03.CNV_res_clonal.rds"}

\item{segdata}{list of clonal_RelativeRatio_segment}
}
\description{
mergeClones()
}
