% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_bulkCNV_estimation.R
\name{mse_loss}
\alias{mse_loss}
\title{mse_loss()}
\usage{
mse_loss(
  seg.cnv.df,
  y_true.term = NULL,
  y_pred.term = NULL,
  weighted = TRUE,
  weight.term = NULL
)
}
\arguments{
\item{seg.cnv.df}{segName  SegMean ratio_map}
}
\description{
the Mean Squared Error (MSE) of integer CNV estimation to the mapping ratio
}
