% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_utils.R
\name{mt2ratio_by_col}
\alias{mt2ratio_by_col}
\title{mt2ratio_by_col()}
\usage{
mt2ratio_by_col(mat, col_idx = NULL, ref.value = NULL)
}
\arguments{
\item{mat}{count maxtrix with cell on column, and feature on row}

\item{col_idx}{integer vector of column start index and end index}

\item{ref.value}{numeric data.frame with the length of nrow(mat)}
}
\description{
calculate bulk-level ratio from single-cell count matrix based on global mean
}
