% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{optimalPloidy}
\alias{optimalPloidy}
\title{optimalPloidy()}
\usage{
optimalPloidy(clusterEst)
}
\arguments{
\item{clusterEst:}{the corrected CN estimation}
}
\value{
the likelihood for each ploidy estimation
}
\description{
ploidy estimation selecting
}
