% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_atac2cnv.R
\name{pcaClustCN}
\alias{pcaClustCN}
\title{pcaClustCN()}
\usage{
pcaClustCN(
  mtx_qcn_lim,
  cell_anno,
  ref_group_names,
  chrs = NULL,
  ref_cells = NULL,
  delimit = "-",
  cytoBandFile = NULL,
  binSize = NULL,
  ClustOnRatio = TRUE,
  seu_resolution = 1,
  min_cells_in_group = 10,
  OutBinMat = TRUE,
  outputFigures = TRUE,
  heatmap_name = NULL,
  count_lim = NULL,
  Figure.width = 10,
  Figure.height = 6,
  FigureName_prefix = NULL,
  FiltSeg = TRUE,
  segRatio_on_bin = TRUE,
  SegLen_min = 1e+06,
  SegSize_min = 5,
  seg_method = "PELT",
  segValue_method = "density",
  seg.count.lim = 80,
  penalty = 1,
  CorlorDot4Ratio = TRUE,
  outdir = ".",
  optimalClust = FALSE,
  mse.min = 0.01,
  cnv.ref.df = NULL,
  global.norm = FALSE
)
}
\arguments{
\item{cnv.ref.df}{data.frame of reference CNV and ratio.map if 'optimalClust' is TRUE}
}
\value{
list of sub-cluster level segmentation, ratio, and CNVs
}
\description{
Perform (1) convert peak to bin matrix if customized required;
(2) Calculate single-cell level ratio relative to bulk normal reference;
(3) PCA clustering based on whole genome bin-cell ratio matrix;
(4) Calculate sub-cluster level CN ratio
}
