% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{peakIndex}
\alias{peakIndex}
\title{peakIndex()}
\usage{
peakIndex(df_seg_C1, initialCN, index_col = "SegMean")
}
\arguments{
\item{df_seg_C1:}{segmentation data}

\item{initialCN:}{the estimation of ratio baseline}
}
\value{
possible ratio vale corresponding to integerCN
}
\description{
finding peak of ration distribution
}
