% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{PloidyCorrect}
\alias{PloidyCorrect}
\title{PloidyCorrect()}
\usage{
PloidyCorrect(integerCNV, delt.lim = 0.3)
}
\arguments{
\item{integerCNV:}{the initial estimation from ABSOLUTE model}
}
\value{
corrected estimation
}
\description{
the relative ratio baseline and distance of integerCN states correction
}
