% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{ploidyEst}
\alias{ploidyEst}
\title{ploidyEst()}
\usage{
ploidyEst(CNest, df_seg_C1)
}
\arguments{
\item{CNest:}{the final integerCN baseline}

\item{df_seg_C1:}{segmentation data}
}
\value{
whole genome ploidy
}
\description{
the overall ploidy of tumor clone
}
