% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{ploidyRefine}
\alias{ploidyRefine}
\title{ploidyRefine()}
\usage{
ploidyRefine(sampelres, delt.lim = 0.3, minCN.frac = 0.01)
}
\arguments{
\item{sampleres}{initial estimation list including all clusters:
$CNVestimate: initial CNV estimation parameters
$seg.dat: segmentatio data
$input_BinSegRatio: bin data}
}
\value{
for the input cluster output: $input_BinSegRatio
$seg.dat
$CNest: ratio corresponding to integerCN
$ploidy: the overall ploidy
}
\description{
the final estimation : select the optimal ploidy estimation for each cluster and correct the negative integer CN
}
