% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{ploidyRefine.ref}
\alias{ploidyRefine.ref}
\title{ploidyRefine.ref()}
\usage{
ploidyRefine.ref(sampelres, CNest.ref, minCN.frac = 0.01, seg_dat_ref = NULL)
}
\arguments{
\item{sampleres}{initial estimation list including all clusters:
$CNVestimate: initial CNV estimation parameters
$seg.dat: segmentatio data
$input_BinSegRatio: bin data}
}
\value{
for the input cluster output: $input_BinSegRatio
$seg.dat
$CNest: ratio corresponding to integerCN
$ploidy: the overall ploidy
}
\description{
the final estimation : estimate clonal CNV based on reference ploidy
}
