% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figurePlot.R
\name{plot4Count}
\alias{plot4Count}
\title{plot4Count()}
\usage{
plot4Count(
  matrix,
  outdir = "./",
  outfile = "rawdataCheck.pdf",
  addline = FALSE,
  format = "pdf",
  min_Nreads_per_cell = NULL,
  min_Nfeature_per_cell = NULL,
  min_Nreads_per_feature = NULL,
  min_Ncell_per_feature = NULL
)

plot4Count(
  matrix,
  outdir = "./",
  outfile = "rawdataCheck.pdf",
  addline = FALSE,
  format = "pdf",
  min_Nreads_per_cell = NULL,
  min_Nfeature_per_cell = NULL,
  min_Nreads_per_feature = NULL,
  min_Ncell_per_feature = NULL
)
}
\description{
Data visualization for the # of reads(or peaks) per cell, and # of reads (or cells) per peak (column : cell)

Data visualization for the # of reads(or peaks) per cell, and # of reads (or cells) per peak (column : cell)
}
