% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figurePlot.R
\name{plot_genome}
\alias{plot_genome}
\title{plot_genome()}
\usage{
plot_genome(
  chr_bg_color = NULL,
  show_x_axis = TRUE,
  special_region = NULL,
  linewidth = 0.5,
  vline = T,
  exclude = c("chrX", "chrY", "chrM")
)
}
\description{
plot summed count of CNAs (amplifications or deletions)
}
