% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_plot.R
\name{plot_point}
\alias{plot_point}
\title{plot_point()}
\usage{
plot_point(
  df,
  x_aes,
  y_aes,
  colour_item = NULL,
  size_item = NULL,
  size_lab = size_item,
  shape_item = NULL,
  shape_label = NULL,
  color_value,
  color_gradient = FALSE,
  color_gradientn_name = "CN",
  color_gradientn_limits = c(1, 6),
  x_lab,
  y_lab,
  label_term = NULL,
  gtitle,
  log_x = FALSE,
  log_y = FALSE,
  legend_position = "rignt",
  legend_color = NULL,
  lm_test = TRUE,
  outfile = NULL,
  out_width = 6,
  out_height = 5,
  xlim.set = TRUE,
  ylim.set = TRUE
)
}
\description{
plot_point()
}
