% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figurePlot.R
\name{plot_segment}
\alias{plot_segment}
\title{plot_segment()}
\usage{
plot_segment(y_value, seg_value, ylab, main, ylim, chromnum, filename)
}
\arguments{
\item{y_value}{for point plot}

\item{seg_value}{for line}

\item{main}{main=paste("Segmentation: ",file_seg,", sub_",sub_cj)}

\item{chromnum}{chromosome annotation with the same length of y_value}

\item{filename}{output paste0(plotDir,"/",de_noise_lb,"segment_",file_seg,"\emph{sub}",sub_cj,".png")}
}
\description{
plot segmentation across genome
}
