% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_pseudoBulk_mat.R
\name{pseudo_bulk_v2}
\alias{pseudo_bulk_v2}
\title{pseudo_bulk_v2()}
\usage{
pseudo_bulk_v2(mat_obj, group_ano, method = "sum", adjust_zero = TRUE)
}
\arguments{
\item{mat_obj}{matrix with features as rownames, cell Id as colnames}

\item{group_ano}{a data.frame of cell_group needs to be provided, the first column is cell ID, the second column is cell groups}

\item{type}{sum or median expression of cells in a group c("mean","sum","median","density")}
}
\description{
segment by computing the changepoint for an ordered vector
(e.g., peaks or bins across chromosome p arm and q arm location)
}
