% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_find_subclones.R
\name{reClustByGroup}
\alias{reClustByGroup}
\title{reClustByGroup()}
\usage{
reClustByGroup(
  mat,
  cell_info,
  cells = NULL,
  resolution = 0.8,
  sep_by = c("-", "-"),
  assay = "binCount"
)
}
\description{
reclustering for subset of matrix by specific cells, and update cell_info in the first column.
cell_info:
subCluster
AAACTCGGTCAGTGTT-1          0
AAACTCGTCTCGACAA-1          0
AAACTGCAGTTATGAG-1          0
}
