% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{refine_segCN.Bayes}
\alias{refine_segCN.Bayes}
\title{refine_segCN.Bayes()}
\usage{
refine_segCN.Bayes(clonalres, seg_dat_ref)
}
\arguments{
\item{clonalres}{A list of clonal segCN results.}

\item{seg_dat_ref}{A data frame of segmented data with columns "ratio" and "integerCN".}
}
\value{
A list of refined segCN results.
}
\description{
This function refines the segCN results using a Bayesian classifier.
}
