% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{refine_segCN.dist}
\alias{refine_segCN.dist}
\title{refine_segCN.dist()}
\usage{
refine_segCN.dist(clonalres)
}
\arguments{
\item{clonalres}{A list of clonal segCN results.}
}
\value{
A list of refined segCN results.
}
\description{
This function refines segCN based on distance to neighbour segments.
}
