% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.subclonality.function.R
\name{regionMerge}
\alias{regionMerge}
\title{regionMerge()}
\usage{
regionMerge(cluster.res)
}
\arguments{
\item{cluster.res:}{output of regionSPlit from all initial bin region}
}
\description{
merg region split result
}
