% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.subclonality.function.R
\name{regionSPlit}
\alias{regionSPlit}
\title{regionSPlit()}
\usage{
regionSPlit(regiondata, mu)
}
\arguments{
\item{regiondata:}{the ratio of bin}

\item{mu:}{the expected ratio of modes}
}
\value{
update mu for the new split region
data: the ratio of new split region with new region cluster ID
clusterindex: 0 no further split
}
\description{
check if there is breakpoint for genomic region
}
