% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.subclonality.function.R
\name{relativeRatio.Ref}
\alias{relativeRatio.Ref}
\title{relativeRatio.Ref()}
\usage{
relativeRatio.Ref(chrratio, ref = 1)
}
\arguments{
\item{ratiodata:}{all ratio value frm all clone and genomic binID}
}
\description{
relative rato data
}
