% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figurePlot.R
\name{replotInfercnv}
\alias{replotInfercnv}
\title{rePlot infercnv heatmap}
\usage{
replotInfercnv(
  expr.mat,
  gene.order.file,
  metadata = NULL,
  left_anno_by = NULL,
  row_split_by = NULL,
  names = "Heatmap",
  left_anno_col = NULL,
  clust_rows = F,
  draw_normal = FALSE,
  ref.mat = NULL,
  width = 10,
  height = 5,
  show_legend_row = TRUE,
  max.legend.value = NULL,
  clustering_method_rows = "complete",
  show_row_names = FALSE,
  direction = "vertical",
  titles = "infercnv Ratio",
  column.title = "Genomic Region",
  legend_side = "right",
  color.breaks = NULL,
  plotDir = "./"
)
}
\arguments{
\item{gene.order.file}{path of gene.order.file}

\item{metadata}{metsdata od cell annotation (cell order)}

\item{names}{output name}

\item{ref.mat}{data.frame or matrix of normal cell cnv (if draw_normal=TRUE)}

\item{expr.file}{data.frame or matrix of tumor cell cnv (row:gene, column:cell)}
}
\description{
rePlot infercnv heatmap
}
