% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeaCNV_main.R
\name{runTeaCNV}
\alias{runTeaCNV}
\title{runTeaCNV()}
\usage{
runTeaCNV(
  input_obj,
  outdir = NULL,
  output_obj_name = "TeaCNV.obj",
  cellFilt_ZeroProp = 0.6,
  correctTn5Batch = FALSE,
  corrected_group_by = NULL,
  sampleID_normal = NULL,
  batchTag = "raw",
  global.norm = FALSE,
  binSize = 1,
  cytoBandFile = NULL,
  seg_method = "PELT",
  segValue.method = "median",
  penalty.lim = c(0.5, 1.5),
  seg.count.lim = 80,
  CorlorDot4Ratio = TRUE,
  FiltSeg = TRUE,
  SegLen_min = 2e+06,
  SegSize_min = 20,
  Zscore_cutoff = 1.28,
  ClustOnRatio = TRUE,
  OutBinMat = TRUE,
  seu_resolution = 1,
  outputFigures = TRUE,
  min_cells_in_group = 20,
  ChrRemove = c("chrX", "chrY", "chrM"),
  numCores = 6,
  minCN_frac = 0.01,
  delt_lim = 0.25,
  StopStep = 4,
  choice = ""
)
}
\arguments{
\item{input_obj}{the TeaCNV object or matrix of peaks (rows) vs. cells (columns) containing the raw counts
It can be a filename or the data directly.}

\item{seg_method}{method for segmentation, including "robseg" and "gaussian".}

\item{annotationFile}{annotation of cells, tab-delimited.}

\item{penalty}{paramter if seg_method is "gaussian"}
}
\description{
runTeaCNV()
}
