% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_find_subclones.R
\name{seg4clusters}
\alias{seg4clusters}
\title{seg4clusters()}
\usage{
seg4clusters(
  mat,
  cell_info,
  cytoBand = NULL,
  col_ref = NULL,
  method = "mean",
  adjust_zero = TRUE,
  seg.penalty = 1,
  seg.method = "PELT",
  segSize.min = 10,
  doFilt = TRUE,
  outplot_path = "./",
  outname = NULL,
  outFigure = TRUE,
  color_seg_gradient = TRUE,
  relative_method = "ratio",
  plot.ylim = NULL
)
}
\arguments{
\item{mat}{single-cell matrix with cells on columns for generate cluster-level pseudo-bulk matrix}

\item{cell_info}{data.frame with cells as rownames, and clusters on the first column}

\item{relative_method}{ratio or Subtraction for relative value}
}
\description{
segmentation for clusters
}
