% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_find_subclones.R
\name{seg_assess}
\alias{seg_assess}
\title{seg_assess()}
\usage{
seg_assess(
  seg_ls,
  sd.cutoff = NULL,
  outname = paste0("./hist.seg_segRatio.pdf"),
  outplot = TRUE
)
}
\arguments{
\item{seg_ls}{list output from seg4clusters()}
}
\description{
seg_assess()
}
