% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_plot.R
\name{seg_compare}
\alias{seg_compare}
\title{seg_compare()}
\usage{
seg_compare(
  df1,
  df2,
  genome = "hg38",
  ylim = NULL,
  ylab = "Relative CN",
  ggtitle = "",
  hline = c(1, 2),
  seg_col = c("darkgrey", "red"),
  alpha = 0.5
)
}
\description{
seg_compare()
}
