% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{seg_filter}
\alias{seg_filter}
\title{seg_filter()}
\usage{
seg_filter(CNVdata, min_bins_per_seg = 1)
}
\arguments{
\item{CNVdata}{a data.frame with segID or breakpoint in column}
}
\description{
filter out the segment with low number of bins (min_bins_per_seg >1)
}
