% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_plot.R
\name{seg_overlap_prop}
\alias{seg_overlap_prop}
\title{seg_overlap_prop()}
\usage{
seg_overlap_prop(region1, region2, split_by = "_|-|:")
}
\arguments{
\item{region1}{and region2, genomic region with format "chrx-xxx-xxx"}
}
\description{
get the percentage of overlap in each region
}
