% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_plot.R
\name{seg_plot}
\alias{seg_plot}
\title{seg_plot()}
\usage{
seg_plot(
  data.bin,
  name.data = "",
  num_cells = NULL,
  plotDir = NULL,
  show_specific_seg = NULL,
  specific_seg_color = "green",
  value.bin = "binRatio",
  value.segment = "SegMean",
  ylab = "Relative CN",
  true_loc = TRUE,
  genome = "hg38",
  plot_seg = TRUE,
  seg_col = "#a52a2a",
  plot_hist = TRUE,
  plot_dots = TRUE,
  log_ratio = FALSE,
  color_dot = FALSE,
  color_hist = TRUE,
  color_hist_gradient = TRUE,
  plot_colors = NULL,
  color_limit = c(0, 3),
  ylim = NULL,
  add_yline = NULL,
  label_gene = NULL,
  label_gene_color = "blue",
  annotations = NULL,
  height = 3,
  width = 14,
  outPlot = TRUE,
  device = "png",
  color_seg_gradient = FALSE,
  disconnected = TRUE
)
}
\arguments{
\item{data.bin}{data.frame of bin level segment with chromosome location as rownames (rowname is important!), such as 'chr1_237513_1530360'}

\item{true_loc}{plot segments and dots at true genomic location. If not true, plot by order.}
}
\description{
segment Plot (on absolute genomic position)
}
