% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_plot.R
\name{seg_plot_dot}
\alias{seg_plot_dot}
\title{seg_plot_dot()}
\usage{
seg_plot_dot(
  data.bin,
  name.data = "",
  plotDir = NULL,
  ylab = "Count",
  genome = "hg38",
  outplot = TRUE
)
}
\description{
seg_plot_dot()
}
