% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_score.R
\name{seg_score}
\alias{seg_score}
\title{seg_score()}
\usage{
seg_score(data, outname = NA, outdir = "./", method = "density")
}
\arguments{
\item{data}{an list output from function 'segment_by_chr_location()', which contains 'changepoint_matrix', 'data_matrix',
and 'Ncells_groups'. The rownames of matrix are 'chr1_237513_1530360' format.
@param method is "mean", "median" or "density". For "density", we will take the density$x from the peak of density as the segMean.
@return res list of two data.frame: segment score at bin-level and segment-level}
}
\description{
score each segment based on the segment length (bin size within the segment),
and the standard deviation of bin values in this segment
}
