% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_plot.R
\name{seg_slimByAlign}
\alias{seg_slimByAlign}
\title{seg_slimByAlign()}
\usage{
seg_slimByAlign(
  segs,
  split_by = "_|-|:",
  Start_site_error = 1e+06,
  End_site_error = 1e+06
)
}
\arguments{
\item{Start_site_error}{the start site distance for same segment (bp)}

\item{End_site_error}{}
}
\description{
Align and unify highly similar fragments
}
