% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_plot.R
\name{seg_split}
\alias{seg_split}
\title{seg_split()}
\usage{
seg_split(segs_input, binSet = NULL, split_by = "_|-|:")
}
\arguments{
\item{binSet}{data.frame with bins in rowname
binID Chromosome   Start     End}
}
\description{
split intersecting segments
}
