% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.correction.v2.R
\name{SegFrac}
\alias{SegFrac}
\title{SegFrac()}
\usage{
SegFrac(df_seg_C1, initialCN, by_term = "SegMean")
}
\arguments{
\item{df_seg_C1:}{segmentation data}

\item{initialCN:}{the initial estimation from ABSOLUTE model}
}
\value{
the fraction of genome
}
\description{
estimating explained genome fraction based on the integerCN estimation used in PloidyCorrect
}
