% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_find_subclones.R
\name{segment4df}
\alias{segment4df}
\title{segment4df()
segmentation per chromosome}
\usage{
segment4df(
  df,
  cytoBand = NULL,
  win = 0,
  outpath = "./",
  doFilt = TRUE,
  segLen_min = 1e+06,
  segSize_min = 5,
  outPlot = TRUE,
  genome = "hg38",
  acen_remove = TRUE,
  plot_ylim = NULL,
  color_dot = TRUE,
  filename = "",
  seg.method = "PELT",
  segValue.method = "median",
  seg.count.lim = 80,
  penalty = 1,
  add_yline = NULL,
  color_seg_gradient = FALSE,
  rmNA = TRUE
)
}
\description{
segment4df()
segmentation per chromosome
}
