% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_find_subclones.R
\name{segment4df_v1}
\alias{segment4df_v1}
\title{segment4df_v1()}
\usage{
segment4df_v1(
  df,
  cytoBand = NULL,
  win = 0,
  outpath = "./",
  doFilt = TRUE,
  segLen_min = 1e+06,
  segSize_min = 5,
  outPlot = TRUE,
  genome = "hg38",
  acen_remove = TRUE,
  plot_ylim = NULL,
  color_dot = TRUE,
  filename = "",
  seg.method = "robseg",
  segValue.method = "density",
  seg.count.lim = 80,
  penalty = 1,
  add_yline = NULL
)
}
\arguments{
\item{df}{data.frame with cluster on column, peak or bin on row}

\item{segValue.method}{the method of take the value of segment based on peaks on it.
segValue.method is "mean", "median" or "density". For "density", we will take the density$x from the peak of density as the segMean.}

\item{seg.count.lim}{the maximum number of segments on whole genome. Default is 80.
cytoBand <- "/Volumes/B/reference/cytoBand_hg38.tsv"}

\item{reference}{file path for segmentation if acen_remove=TRUE}
}
\description{
segmentation for each column (cluster) of data.frame
}
