% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figurePlot.R
\name{segPlot4CNV}
\alias{segPlot4CNV}
\title{segPlot4CNV()
plot segment with estimated CNV}
\usage{
segPlot4CNV(CNVres, integerCNV, outdir, fname)
}
\arguments{
\item{integerCNV}{a vector of integerCNV with the same length of CNVres.}
}
\description{
segPlot4CNV()
plot segment with estimated CNV
}
