% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{smooth_and_denoise}
\alias{smooth_and_denoise}
\title{smooth_and_denoise()}
\usage{
smooth_and_denoise(mat_obj, window = 10, sd_amplifier = 1)
}
\arguments{
\item{mat_obj}{cells in columns}
}
\description{
smooth_and_denoise()
}
