% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{split_grange}
\alias{split_grange}
\title{split_grange()}
\usage{
split_grange(bins_input, cytoBand = NULL)
}
\arguments{
\item{bins_input}{strings of genomic regions,"chrx-xxx-xxx" format}

\item{cytoBand}{the reference annotation of chromosome location, BED format}
}
\value{
bed, with the fourth column is bins_input, the second column is new bin ID, which
breaks by acen
}
\description{
split_grange()
}
