% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_segmentation_comb.R
\name{strings2bed}
\alias{strings2bed}
\title{strings2bed()}
\usage{
strings2bed(strings, split_by = "_|-|:", check.sort = TRUE)
}
\description{
convert strings (e.g., "chrx-xxx-xxx") to bed format
}
