% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydataProcess.R
\name{symbol2Coordinate}
\alias{symbol2Coordinate}
\title{map2chrLocatio()}
\usage{
symbol2Coordinate(vect, refFile)
}
\arguments{
\item{refFile}{Bed format in first three columns "chr" "start" "end"}
}
\description{
convert gene symbol to gnomic Coordinate, vecter format
}
